Update main.js
When empty cells are merged, hLen and vLen may not get proper values, which leads to collapsed borders in some circumstances.

Example: Put some contents in cells (0, 0) and (2, 0) (where (2,0) means row 2 and column 0, etc). Then merge empty cells (1,0) and (1,1). The collapsed right border is visible when choosing single or double vertical borders.

The solution is simple: in getHLen and getVLen, include values from merged cells even if they are empty. I opted to tag the merged cells, but one could also check if either rowspan or colspan is greater than 1.