Update index.html
Another bug about merged cells: put content in cell (0,0), then merge with cell (1,0) below. The select delete all the rows except the first one (ie select the rows from 1 to the end below, so all rows except the first one, and click 'delete row' in the context menu to delete these rows. These triggers some error in `handsontable` (see the console while doing this).

This seems to be a `handsontable` bug for this old version. Haven't tried which immediate subsequent version solves the problem, but the bug is not present when using `handsontable` v 6.2.2, which is the last version with MIT license (for the Community Edition).

`main.js` also has to be modified to properly register the `mergeCells` plugin.